+++ inttest/codegeneration-defaults/model.cpp
@@ -57,7 +57,7 @@ Type objective_function<Type>::operator() () {
     PARAMETER(fish__Linf);
     PARAMETER(fish__K);
     PARAMETER(fish__t0);
-    PARAMETER(fish__init__sd);
+    PARAMETER(fish__lencv);
     PARAMETER(fish__walpha);
     PARAMETER(fish__wbeta);
     PARAMETER(report_detail);
@@ -363,7 +363,7 @@ Type objective_function<Type>::operator() () {
 
                     auto factor = (fish__init__scalar*map_extras::at_throw(fish__init, std::make_tuple(age), "fish.init")*exp(-(double)(1)*(fish__M + init__F)*(age - recage)));
 
-                    auto dnorm = ((fish__midlen - (fish__Linf*((double)(1) - exp(-(double)(1)*fish__K*((age - cur_step_size) - fish__t0))))) / fish__init__sd);
+                    auto dnorm = ((fish__midlen - (fish__Linf*((double)(1) - exp(-(double)(1)*fish__K*((age - cur_step_size) - fish__t0))))) / ((fish__Linf*((double)(1) - exp(-(double)(1)*fish__K*(age - fish__t0))))*fish__lencv));
 
                     {
                         fish__num.col(fish__age_idx).col(fish__area_idx) = normalize_vec(exp(-(pow(dnorm, (Type)(double)(2)))*(double)(0.5)))*(double)(10000)*factor;