+++ inttest/codegeneration-defaults/run.R
@@ -23,7 +23,7 @@ actions_fish <- list(
   g3a_growmature(fish, g3a_grow_impl_bbinom(
     maxlengthgroupgrowth = 5L)),
   g3a_naturalmortality(fish),
-  g3a_initialconditions_normalparam(fish),
+  g3a_initialconditions_normalcv(fish),
   g3a_renewal_normalparam(fish),
   g3a_age(fish),
   NULL)