+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## New features
 * Add ``g3_parameterized('x', by_year = 1998:2099)`` to override year range for parameters
+* Add ``g3a_initialconditions_normalcv`` for initialconditions driven by a ``lencv`` parameter
 
 ## Breaking changes